ID-1393, ID-1386: Finalization of GraphScope configuration
* Remove all parameters with default values for simplification
* set subGraphLimit to 100
* consistent app names (app-{mode}-default), example query
* comment the graphscope service definition docker-compose.override file
to adhere to the current description in readme (when graphscope should
be removed)

With ID-1386 the GraphScope server now may use environment variables to
set config values. This change prepares the config files to optionally
read remote user and password from environment variables if set. When
unset if falls back to the provided (and unchanged) default values.