Update GraphDB compose environment to GraphDB 10 + refine license info
With this change we make GraphDB 10 the default database. As GraphDB
provides a new license mechanism (incl. a Free Mode), we adjust our
instructions accordingly.

The license is no longer mounted into the container as a volume, but can
be set through the GraphDB workbench. This makes trial mode with GraphDB
as backing database easier.